requirements.txt
asgiref==3.8.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
filelock==3.16.1
fsspec==2024.12.0
jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.0.0
sqlparse==0.5.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing-extensions==4.12.2
tzdata==2024.2
